pyproject.toml
[project]
name = "dnd-dm-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.8.0",
    "jinja2>=3.1.6",
    "litellm>=1.74.8",
    "pydantic>=2.11.7",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.5",
]
